requirements.txt
bilibili-api-python>=17.4.1
anthropic>=0.18.0
httpx>=0.24.0
